Sort files according to language
