+++ C++/armstrong-number.cpp
