+++ Python/Armstrong.py
