+++ Python/spirograph.py
